+++ src/ShowRelease.hs
@@ -10,12 +10,13 @@ module ShowRelease (
 where
 
 import Control.Monad.Extra (forM_, unless, void, when, whenJust)
-import qualified Data.ByteString.Char8 as B
+import Data.Bifunctor (first)
+import qualified Data.CaseInsensitive as CI
 import Data.List.Extra
 import Data.Maybe (fromMaybe)
+import Data.Time.Format (defaultTimeLocale, parseTimeM, rfc822DateFormat)
 import Data.Time.LocalTime (utcToLocalZonedTime)
-import Network.HTTP.Directory (httpLastModified, httpManager,
-                               httpRedirect, Manager)
+import Network.Curl (curlHead, CurlOption(..))
 import SimpleCmd (error', (+-+))
 import Text.Regex (mkRegex, subRegex)
 
@@ -33,11 +34,10 @@ showReleaseCmd debug dynredir reposource release sysarch march testing =
 showRelease :: Bool -> Bool -> Bool -> Bool -> RepoSource
             -> Release -> Arch -> Maybe Arch -> Bool -> IO [(String, URL)]
 showRelease debug dynredir warn checkdate reposource@(RepoSource koji _chan _mirror) release sysarch march testing = do
-  mgr <- httpManager
   let arch = fromMaybe sysarch march
-  (url,path) <- getURL debug dynredir mgr reposource release arch
+  (url,path) <- getURL debug dynredir reposource release arch
   let urlpath = url +//+ path
-  when debug $ print $ renderUrl urlpath
+  when debug $ print $ renderUrl Dir urlpath
   (basicrepos,morerepos) <-
     case release of
       -- RepoKoji -> ["koji-fedora"]
@@ -68,36 +68,37 @@ showRelease debug dynredir warn checkdate reposource@(RepoSource koji _chan _mir
         map (repoConfigArgs reposource sysarch march rawhide release) morerepos
   forM_ basicrepourls $ \(reponame,(url',path')) -> do
     let baserepo = url' +//+ path'
-    when debug $ print $ renderUrl baserepo
+    when debug $ print $ renderUrl Dir baserepo
     unless (not checkdate || release == System) $ do
       let composeinfo =
+            url' +//+
             if koji
-            then url' +//+ ["repo.json"]
+            then ["repo.json"]
             else
               case release of
-                Centos 10 -> url' +//+ ["metadata","composeinfo.json"]
-                Centos _ -> url' +//+ ["COMPOSE_ID"] -- ["metadata","composeinfo.json"]
-                ELN -> url' +//+ ["metadata","composeinfo.json"]
-                EPEL _ -> url' +//+ ["Everything", "state"]
-                EPELNext _ -> url' +//+ ["Everything", "state"]
-                Fedora _ -> url' +//+
-                            if "updates" `isInfixOf` reponame ||
+                Centos 10 -> ["metadata","composeinfo.json"]
+                Centos _ -> ["COMPOSE_ID"] -- ["metadata","composeinfo.json"]
+                ELN -> ["metadata","composeinfo.json"]
+                EPEL _ -> ["Everything", "state"]
+                EPELNext _ -> ["Everything", "state"]
+                Fedora _ -> if "updates" `isInfixOf` reponame ||
                                "updates-testing" `isInfixOf` reponame
                             then ["Everything", "state"]
                             else ["COMPOSE_ID"]
-                Rawhide -> url' +//+ ["COMPOSE_ID"]
+                Rawhide -> ["COMPOSE_ID"]
                 System -> error' "system not supported"
-      let composeUrl = renderUrl composeinfo
+      let composeUrl = renderUrl File composeinfo
       when debug $ print composeUrl
-      mtime <- httpLastModified mgr composeUrl
-      whenJust mtime $ \utc -> do
+      mtimestr <- curlGetHeader "Last-Modified" composeUrl
+      whenJust mtimestr $ \timestr -> do
+        utc <- parseTimeM False defaultTimeLocale rfc822DateFormat timestr
         date <- utcToLocalZonedTime utc
-        (if warn then warning else putStrLn) $ show date +-+ "<" ++ renderUrl url' ++ ">"
+        (if warn then warning else putStrLn) $ show date +-+ "<" ++ renderUrl Dir url' ++ ">"
   return $ map (fmap (uncurry (+//+))) $ basicrepourls ++ morerepourls
 
-getURL :: Bool -> Bool -> Manager -> RepoSource -> Release -> Arch
+getURL :: Bool -> Bool -> RepoSource -> Release -> Arch
        -> IO (URL,[String])
-getURL debug dynredir mgr reposource@(RepoSource koji chan _mirror) release arch =
+getURL debug dynredir reposource@(RepoSource koji chan _mirror) release arch =
   case release of
     Centos n ->
       case n of
@@ -120,8 +121,8 @@ getURL debug dynredir mgr reposource@(RepoSource koji chan _mirror) release arch
     EPEL n | n < 7 ->
                return
                (URL "https://archives.fedoraproject.org/pub/archive/epel", [show n])
-    EPEL n -> getFedoraServer debug dynredir mgr reposource ["epel"] [show n]
-    EPELNext n -> getFedoraServer debug dynredir mgr reposource ["epel","next"] [show n]
+    EPEL n -> getFedoraServer debug dynredir reposource ["epel"] [show n]
+    EPELNext n -> getFedoraServer debug dynredir reposource ["epel","next"] [show n]
     Fedora n -> do
       ebodhirelease <- activeFedoraRelease n
       case ebodhirelease of
@@ -135,9 +136,9 @@ getURL debug dynredir mgr reposource@(RepoSource koji chan _mirror) release arch
           let pending = releaseState rel == "pending"
               rawhide = pending && releaseBranch rel == "rawhide"
               releasestr = if rawhide then "rawhide" else show n
-          in getFedoraServer debug dynredir mgr reposource fedoraTop
+          in getFedoraServer debug dynredir reposource fedoraTop
              [if pending then "development" else "releases", releasestr]
-    Rawhide -> getFedoraServer debug dynredir mgr reposource fedoraTop ["development", "rawhide"]
+    Rawhide -> getFedoraServer debug dynredir reposource fedoraTop ["development", "rawhide"]
     System -> error' "getURL: system unsupported"
   where
     fedoraTop =
@@ -196,24 +197,26 @@ repoConfigArgs (RepoSource True _chan _mirror) sysarch march _rawhide release (r
 --       reponame = repo ++ "-Centos-" ++ show release ++ "-Stream" ++ "-" ++ show chan ++ if arch == sysarch then "" else "-" ++ showArch arch
 --   in (reponame, (url +//+ compose, [path, showArch arch, "os/"]))
 
-getFedoraServer :: Bool -> Bool -> Manager -> RepoSource -> [String] -> [String]
+getFedoraServer :: Bool -> Bool -> RepoSource -> [String] -> [String]
                 -> IO (URL,[String])
-getFedoraServer debug dynredir mgr (RepoSource koji _ mirror) top path =
+getFedoraServer debug dynredir (RepoSource koji _ mirror) top path =
   if koji
   then return (URL "https://kojipkgs.fedoraproject.org",[])
   else
     case mirror of
       DownloadFpo -> do
         let url = URL downloadServer +//+ top ++ path
-            rurl = renderUrl url
+            rurl = renderUrl Dir url
+        when debug $ print rurl
         redir <-
           if dynredir
           then return Nothing
-          else fmap B.unpack <$> httpRedirect mgr rurl
+          else curlGetHeader "location" rurl
+        when debug $ print redir
         case redir of
           Nothing -> do
             unless dynredir $
-              warning $ "no redirect for" +-+ rurl
+              warning $ "slow redirection for" +-+ rurl
             return (URL downloadServer +//+ top,path)
           Just actual -> do
             when debug $ do
@@ -232,3 +235,10 @@ getFedoraServer debug dynredir mgr (RepoSource koji _ mirror) top path =
 
 downloadServer :: String
 downloadServer = "https://download.fedoraproject.org/pub"
+
+-- FIXME check CurlCode(CurlOK) status
+curlGetHeader :: CI.CI String -> String -> IO (Maybe String)
+curlGetHeader field url = do
+  (_status,headers) <- curlHead url [CurlConnectTimeoutMS 5000]
+  -- tail because curl leaves space before field value
+  return $ tail <$> lookup field (map (first CI.mk) headers)